+++ pyaud/__init__.py
@@ -4,7 +4,7 @@ from ._environ import load_namespace
 from ._indexing import files
 from ._objects import BasePlugin
 from ._subprocess import git
-from ._utils import branch, get_packages, package
+from ._utils import branch, get_commit_hash, get_packages, package
 from ._version import __version__
 
 __all__ = [
@@ -14,6 +14,7 @@ __all__ = [
     "config",
     "exceptions",
     "files",
+    "get_commit_hash",
     "get_packages",
     "git",
     "load_namespace",